+++ src/Tables/Filters/DateFilter.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Happones\Kinetix\Tables\Filters;
 
+use Happones\Kinetix\Support\KinetixLocale;
 use Illuminate\Database\Eloquent\Builder;
 
 class DateFilter extends Filter
@@ -92,7 +93,7 @@ class DateFilter extends Filter
     {
         return [
             'useCalendar' => $this->useCalendar,
-            'locale'      => $this->locale,
+            'locale'      => $this->locale ?? KinetixLocale::bcp47(),
             'minValue'    => $this->minValue,
             'maxValue'    => $this->maxValue,
         ];